chore: Move stratCol dialig. Add input validation and error feedback.
